requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
crispy-bootstrap5==2024.10
cryptography==44.0.0
daphne==4.1.2
Django==4.2.19
django-allauth==65.4.1
django-axes==7.0.1
django-countries==7.6.1
django-crispy-forms==2.3
django-recaptcha==4.0.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
jwt==1.3.1
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-decouple==3.8
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
service-identity==24.2.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zope.interface==7.2